# swaggerv2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Ecommerce RESTful API",
    "version": "2.0.0",
    "description": "A backend RESTful API for an Ecommerce Store",
    "contact": {
      "name": "Custom Haven",
      "url": "https://github.com/CustomHaven"
    },
    "license": {
      "name": "MIT",
      "url": "https://choosealicense.com/licenses/mit/"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4001/",
      "variables": {}
    },
    {
      "url": "https://rest-api-no-orm.herokuapp.com/",
      "variables": {}
    }
  ],
  "paths": {
...
    "/v2/contact-details": {
      "summary": "User's shipping details",
      "description": "Where a user lives so their items can be shipped to them",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "All shipping address",
        "description": "All shipping details can only be viewed by an admin!",
        "operationId": "fetch_all_shipping_details",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "[{\"id\": 1, \"user_id\": \"user 1\", \"address_line1\": \"05412 Maywood Plaza\", \"address_line2\": \"0 Talmadge Court\", \"town_city\": \"Conceição do Jacuípe\", \"zip_code\": \"44245-000\", \"country\": \"Brazil\", \"created\": \"10/11/2021\", \"modified\": \"22/07/2022\" }, { \"id\": 2, \"user_id\": \"user 2\",\"address_line1\": \"18 Gulseth Street\", \"address_line2\": \"8051 Prairie Rose Avenue\", \"town_city\": \"Furukawa\", \"zip_code\": \"963-8823\",   \"country\": \"Japan\", \"created\": \"05/10/2021\", \"modified\": \"16/03/2022\" }]"
                  }
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": [
          "Contact Details"
        ]
      }
    },
    "/v2/contact-details/{user_id}": {
      "summary": "User's shipping details",
      "description": "Where a user lives so their items can be shipped to them",
      "post": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Create shipping details",
        "description": "Setting up the Shipping details for the specific user",
        "operationId": "registering_shipping_details",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "The user's id from the Users table",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact-Details"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"id\": 3, \"user_id\": \"user 3\", \"address_line1\": \"44347 Nobel Place\", \"address_line2\": \"02947 Warner Hill\", \"town_city\": \"Baddomalhi\",   \"zip_code\": \"22501\", \"country\": \"Pakistan\",\"created\": \"14/10/2021\", \"modified\": \"21/02/2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          }
        },
        "tags": [
          "Contact Details"
        ]
      }
    },
    "/v2/contact-details/{id}": {
      "summary": "Specific user's details",
      "description": "Fetch and do something with a user's contact details based on the given parameter",
      "get": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Display shipping details",
        "description": "Display a user's shipping details to them",
        "operationId": "display_shipping_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the contact detail",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\"id\": 1, \"user_id\": \"user 1\", \"address_line1\": \"05412 Maywood Plaza\", \"address_line2\": \"0 Talmadge Court\", \"town_city\": \"Conceição do Jacuípe\", \"zip_code\": \"44245-000\", \"country\": \"Brazil\", \"created\": \"10/11/2021\", \"modified\": \"22/07/2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{ \"message\": \"Failed to fetch details\" }"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Contact Details"
        ]
      },
      "put": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Update a user's contact details",
        "description": "Display a user's contact details to them",
        "operationId": "update_contact_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the contact detail",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "description": "Sign up a new user",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Contact-Details"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{\"id\": 1, \"user_id\": \"user 1\", \"address_line1\": \"10 Downing Street\", \"address_line2\": \"Westminister\", \"town_city\": \"London\", \"zip_code\": \"SW1A 2AA\", \"country\": \"United Kingdom\", \"created\": \"10/11/2021\", \"modified\": \"22/07/2022\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{ \"message\": \"Failed to fetch details\" }"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Contact Details"
        ]
      },
      "delete": {
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "summary": "Delete shipping detail",
        "description": "Delete a user's shipping details",
        "operationId": "delete_shipping_details",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "The id of the contact detail",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "content": {
              "application/json": {
                "examples": {
                  "success": {
                    "value": "{ \"message\": \"Successfully deleted\" }"
                  }
                }
              }
            },
            "description": "Success"
          },
          "404": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{ \"message\": \"Failed to fetch details\" }"
                  }
                }
              }
            },
            "description": "Fail"
          },
          "500": {
            "content": {
              "application/json": {
                "examples": {
                  "fail": {
                    "value": "{\n  \"message\": \"Invalid Server Error\"\n}"
                  }
                }
              }
            },
            "description": "Fail"
          }
        },
        "tags": [
          "Contact Details"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Contact-Details": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "user_id": {
            "type": "string"
          },
          "address_line1": {
            "type": "string"
          },
          "address_line2": {
            "type": "string"
          },
          "town_city": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "created": {
            "type": "string"
          },
          "modified": {
            "type": "string"
          }
        },
        "xml": {
          "name": "Contact-Details"
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect_sid"
      }
    }
  }
}